backend/requirements.txt
fastapi
uvicorn
gunicorn
pypdf
numpy
google-genai
python-dotenv
langchain-text-splitters
tiktoken
umap-learn